requirements.txt
brainflow>=5.0.0
PyQt5>=5.15.0
pyqtgraph>=0.13.0
numpy>=1.21.0,<2.0.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0
pyserial>=3.5
